Add Gemfile and puppet 4 checks
In anticipation of puppet 4, start trying to deal with puppet 4 things
that can be helpfully predicted by puppet lint plugins. Also fix lint
errors caught by the puppet-lint-absolute_classname-check gem and arrow
alignment errors now caught by the modern puppet-lint gem.

We disable the puppet-lint-empty_string-check check for now, to be
fixed later, because this module makes heavy use of empty string
defaults and updating this needs to be done with care.

Change-Id: I90e74f25a1aa176d336afc591ac1da9c03c77aa7