php5-fpm 503 errors
formerly we were using settings for php5-fpm pool that
werent suitable for our server configuration.
Server has 14G of free ram and we were using
pm.max_children = 4096, having in consideration that php process
has an avg of 64 MB that give us large number of needed
RAM that exceeds our current configuration
i changed to pm.max_children = 200, a more suitable number
for our current config
rite now the avg php if given us 25 MB per process
so i guess that we should do a monitoring of this value, an
increase the max_children setting as we needeed on the fly.

Change-Id: Ibab02f7a88e08cb94d58168e545e6da0cf8d2e7f